fix(tests): MacroDistributionDto float -> int + corrections assertions #NTR-9
- MacroDistributionDto : champs float -> int (ProteinPct, CarbPct, FatPct)
- Correction instanciations dans AdminControllerTest, DietPlansControllerTest, DietsControllerTest
- CLAUDE.md : mise à jour état du projet

+++ src/NutritionApi.Application/DTOs/DietPlans/MacroDistributionDto.cs
@@ -1,8 +1,8 @@
-
-using NutritionApi.Domain.ValueObjects;
 namespace NutritionApi.Application.DTOS.DietPlans;
 
-public record MacroDistributionDto(float ProteinPct, float CarbPct, float FatPct)
+using NutritionApi.Domain.ValueObjects;
+
+public record MacroDistributionDto(int ProteinPct, int CarbPct, int FatPct)
 {
     public static MacroDistributionDto From(MacroDistribution macro)
             => new(macro.ProteinPercentage, macro.CarbPercentage, macro.FatPercentage);
